Fix linting issues with DatasetSorter components.
