+++ packages/openneuro-app/src/scripts/datalad/dashboard/datasets/dataset-sorter.jsx
@@ -1,0 +1,72 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import styled from 'styled-components'
+
+// DatasetSort GraphQL fields
+const sortFields = [
+  'created',
+  'name',
+  'uploader',
+  'stars',
+  'downloads',
+  'subscriptions',
+]
+
+const Capitalized = styled.span`
+  text-transform: capitalize;
+`
+
+const SortField = ({ field, queryVariables, refetch }) => {
+  const fieldValue =
+    field in queryVariables.orderBy && queryVariables.orderBy[field]
+  let icon
+  if (fieldValue) {
+    if (fieldValue === 'ascending') {
+      icon = <i className="fa fa-sort-asc" />
+    } else {
+      icon = <i className="fa fa-sort-desc" />
+    }
+  }
+  const sortBy = () => {
+    const newQueryVariables = { ...queryVariables }
+    newQueryVariables.orderBy[field] =
+      queryVariables.orderBy[field] === 'descending'
+        ? 'ascending'
+        : 'descending'
+    refetch(queryVariables)
+  }
+  return (
+    <a
+      key={field}
+      className={fieldValue ? 'btn-sort name active' : 'btn-sort name'}
+      onClick={sortBy}>
+      <Capitalized>{field}</Capitalized> {icon}
+    </a>
+  )
+}
+
+SortField.propTypes = {
+  field: PropTypes.string,
+  queryVariables: PropTypes.object,
+  refetch: PropTypes.func,
+}
+
+const DatasetSorter = ({ queryVariables, refetch }) => (
+  <>
+    {sortFields.map(field => (
+      <SortField
+        field={field}
+        queryVariables={queryVariables}
+        refetch={refetch}
+        key={field}
+      />
+    ))}
+  </>
+)
+
+DatasetSorter.propTypes = {
+  queryVariables: PropTypes.object,
+  refetch: PropTypes.func,
+}
+
+export default DatasetSorter
